--- docs/source/guide/reference/mpa/modules/models/index.rst
@@ -1,8 +1,0 @@
-Models
-^^^^^^^
-
-.. toctree::
-   :maxdepth: 1
-
-   losses
-   utils
