--- docs/source/guide/reference/mpa/modules/models/losses.rst
@@ -1,14 +1,0 @@
-Losses
-^^^^^^^
-
-.. toctree::
-   :maxdepth: 3
-   :caption: Contents:
-
-.. automodule:: otx.mpa.modules.models.losses
-   :members:
-   :undoc-members:
-
-.. automodule:: otx.mpa.modules.models.losses.utils
-   :members:
-   :undoc-members:
