--- docs/source/guide/reference/mpa/modules/models/utils.rst
@@ -1,10 +1,0 @@
-Utils
-^^^^^^^
-
-.. toctree::
-   :maxdepth: 3
-   :caption: Contents:
-
-.. automodule:: otx.mpa.modules.models.utils
-   :members:
-   :undoc-members:
