+++ otx/algorithms/common/adapters/mmcv/hooks/task_adapt_hook.py
@@ -6,8 +6,7 @@
 from mmcv.runner import HOOKS, Hook, get_dist_info
 from torch.utils.data import DataLoader
 
-from otx.mpa.modules.datasets.samplers.balanced_sampler import BalancedSampler
-from otx.mpa.modules.datasets.samplers.cls_incr_sampler import ClsIncrSampler
+from otx.algorithms.common.adapters.torch.dataloaders.samplers import BalancedSampler, ClsIncrSampler
 from otx.mpa.utils.logger import get_logger
 
 logger = get_logger()
