+++ otx/algorithms/common/adapters/torch/dataloaders/__init__.py
@@ -3,3 +3,7 @@
 #
 
 # flake8: noqa
+
+from .composed_dataloader import ComposedDL
+
+__all__ = ['ComposedDL']
