+++ tests/unit/algorithms/common/adapters/torch/dataloaders/__init__.py
@@ -1,5 +1,4 @@
+"""Test for otx.algorithms.common.adapters.torch.dataloaders"""
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 #
-
-# flake8: noqa
