--- tests/unit/mpa/modules/datasets/__init__.py
@@ -1,4 +1,0 @@
-"""Test for otx.mpa.modules.datasets"""
-# Copyright (C) 2022 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
